requirements.txt
discord
python-dotenv
appdirs
